requirements.txt
python-dotenv==1.0.1          # Загрузка переменных из .env
pandas==2.2.1                 # Анализ данных, исторические свечи
numpy==1.26.4                 # Вычисления, индикаторы
requests==2.31.0              # HTTP-запросы к API Bybit
websockets==15.0.1            # WebSocket для ордербука, сделок
pybit==5.9.0                  # Альтернативный SDK Bybit (если понадобится)
tabulate==0.9.0               # Красивая печать табличек
pytest==8.0.2                 # Тестирование
requests-mock==1.11.0         # Мокаем API в тестах
pycryptodome==3.21.0          # Подписи HMAC для ордеров
aiologger==0.7.0              # Асинхронные логи
psutil==5.9.8                 # Мониторинг ресурсов
ta==0.10.2                    # Индикаторы (EMA, RSI, etc.)
